requirements.txt
flask
flask_cors
transformers
pillow
torch